SOFB.FIX: Make sure BPMs are properly configured at initialization.
This commit forces the configuration of the proper SOFBmode-Sts to the BPMs and timing system.

This breaks our convention of not setting any PVs during initialization, but I think it is worth it, because nowadays TL SOFBs generally initialize in a broken state and operators forget/don't notice they have to reconfigure the acquisitions, which makes the IOC fail to get new trajectories.